verify_setup.py
import os
import sys

try:
    print("Importing modules...")
    import streamlit
    import google.generativeai
    from PIL import Image
    from dotenv import load_dotenv
    
    from image_gen import generate_assets
    from stt import transcribe_audio
    from agents import JournalistAgent, EvaluatorAgent
    from interview_engine import InterviewSession
    
    print("Modules imported successfully.")
    
    print("Instantiating classes...")
    session = InterviewSession()
    # Note: We can't fully instantiate Agents without a valid API key in env, 
    # but we can check if the class exists and methods are defined.
    
    print("Verification complete. No syntax errors found.")
    
except Exception as e:
    print(f"Verification failed: {e}")
    sys.exit(1)
