requirements.txt
streamlit
google-generativeai
google-genai
Pillow
python-dotenv
